Some cleanup and use ArrayPool.Shared for line buffers (#41)
* Correct namespace and move constants that are only used once.

* Add spelling exclusion file and update unit tests

* ThrowHelper: always throw from a separate method

Throwing from a separate method allows the JIT compiler to only inline the checking method. By default the JIT compiler won't inline methods that have a singel throw statement.

* Move validation methods from util to validation

* Use ArrayPool.Shared for line buffers

Note: the assumption is that the type of the array doesn't matter. In principle we only need byte arrays.

* Change FrameInfo from class into struct

FrameInfo is only 16 bytes, good candidate for a struct.

* Resolve SonarLint warnings

Note: some sonarlint warnings are ignored, but not all can be ignored in the GlobalSuppression file.