Add property JpegLSDecoder.CompressedDataFormat
The JPEG-LS standard defines 3 compressed data formats. Only the interchange format (the most common) can normally be decoded.
Add a method to allow application code to retrieve the data format of the JPEG-LS stream. This allows an application to decide if it want to display the image
or process it differently.